Got entries to show up
Entries show up, but we still need to link each dream to a user for the
query to pull entries belonging to that user. right now i just have it
pulling entries that match to a "test" boolean column in parse.  This
can easily be changed to match a username.

To get everything working, delete the "Parse-1.10.1.jar" in
DreamLink\apps\libs\ because I replaced it "Parse-1.4.1.jar" which has
support for the ParseQueryAdapter and stuff. Then
rebuild/refresh/whatever your gradle. Also, you can get rid of the
"item_layout.xml"

I still need to touch up some stuff and add the onItemClick()
functionality but this is what I have so far  :+1: